+++ js/util.js
@@ -1,0 +1,5 @@
+const getRandomInteger = (min, max) => Math.floor(Math.random() * (Math.floor(max) - Math.ceil(min) + 1) + Math.ceil(min));
+
+const getRandomArrElem = (array) => array[Math.floor(Math.random() * array.length)];
+
+export {getRandomInteger, getRandomArrElem};
